test: add unit tests for CustomLensCore to verify custom method preservation in focus, reflect, and create operations
